requirements.txt
click==8.1.3
Flask==2.2.2
Flask-WTF==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg2-binary==2.9.5
types-psycopg2==2.9.21.4
typing_extensions==4.4.0
Werkzeug==2.2.2
WTForms==3.0.1
